+++ src/bitpacking.rs
@@ -1,7 +1,6 @@
 use arrayref::{array_mut_ref, array_ref};
 use const_for::const_for;
 use core::mem::size_of;
-use num_traits::AsPrimitive;
 use paste::paste;
 
 use crate::{pack, seq_t, unpack, FastLanes, Pred, Satisfied, FL_ORDER};
@@ -57,31 +56,6 @@ pub trait BitPacking: FastLanes {
     /// is the packed width. The output slice must be of exactly length 1024.
     /// These lengths are checked only with `debug_assert` (i.e., not checked on release builds).
     unsafe fn unchecked_unpack_single(width: usize, input: &[Self], index: usize) -> Self;
-}
-
-pub trait BitPackingCompare: BitPacking {
-    fn unpack_cmp_impl<const W: usize, F: Fn(Self, Self) -> bool>(
-        input: &[Self; 1024 * W / Self::T],
-        output: &mut [Self; 1024 / Self::T],
-        f: F,
-        eq_value: Self,
-    ) where
-        BitPackWidth<W>: SupportedBitPackWidth<Self>;
-
-    fn unpack_cmp<const W: usize, F: Fn(Self, Self) -> bool>(
-        input: &[Self; 1024 * W / Self::T],
-        output: &mut [u64; 16],
-        comparison: F,
-        eq_value: Self,
-    ) where
-        BitPackWidth<W>: SupportedBitPackWidth<Self>,
-        [(); 1024 / Self::T]:,
-    {
-        let new_output = unsafe {
-            &mut *std::ptr::from_mut::<[u64; 16]>(output).cast::<[Self; 1024 / Self::T]>()
-        };
-        Self::unpack_cmp_impl(input, new_output, comparison, eq_value);
-    }
 }
 
 macro_rules! impl_packing {
@@ -262,37 +236,6 @@ impl_packing!(u16);
 impl_packing!(u32);
 impl_packing!(u64);
 
-macro_rules! impl_packing_compare {
-    ($T:ty) => {
-        paste! {
-            impl BitPackingCompare for $T {
-               #[inline(always)]
-                fn unpack_cmp_impl<const W: usize, F: Fn(Self, Self) -> bool>(
-                    input: &[Self; 1024 * W / Self::T],
-                    output: &mut [Self; 1024 / Self::T],
-                    f: F,
-                    other: Self,
-                ) where BitPackWidth<W>: SupportedBitPackWidth<Self> {
-                    for lane in (0..Self::LANES){
-                        unpack!($T, W, input, lane, |$idx, $elem| {
-                            let bool_idx = $idx / Self::T;
-                            let bool_bit = $idx % Self::T;
-                            let value = f($elem, other);
-                            output[bool_idx] |= (AsPrimitive::<Self>::as_(value)) << bool_bit;
-                        });
-                    }
-                }
-
-            }
-        }
-    };
-}
-
-// TODO(joe): fix this.
-// Do not impl this for u8/u16 as its currently slower.
-impl_packing_compare!(u32);
-impl_packing_compare!(u64);
-
 #[cfg(test)]
 mod test {
     use core::array;
@@ -370,63 +313,4 @@ mod test {
     seq!(W in 0..=16 { impl_try_round_trip!(u16, W); });
     seq!(W in 0..=32 { impl_try_round_trip!(u32, W); });
     seq!(W in 0..=64 { impl_try_round_trip!(u64, W); });
-
-    #[test]
-    fn test_unpack_eq() {
-        type T = u32;
-        const W: usize = 3;
-
-        let values = array::from_fn(|i| i as T % 32);
-
-        let mut packed = [0; (128 * W) / size_of::<T>()];
-        T::pack::<W>(&values, &mut packed);
-
-        let mut output = [0u64; 16];
-        T::unpack_cmp::<W, _>(&packed, &mut output, |a, b| a == b, 4);
-
-        let mut unpacked = [0; 1024];
-        T::unpack::<W>(&packed, &mut unpacked);
-        let bools = collect_bool(unpacked.len(), |idx| unpacked[idx] == 4);
-
-        assert_eq!(output.as_slice(), bools.as_slice());
-    }
-
-    #[inline]
-    pub fn ceil(value: usize, divisor: usize) -> usize {
-        // Rewrite as `value.div_ceil(&divisor)` after
-        // https://github.com/rust-lang/rust/issues/88581 is merged.
-        value / divisor + usize::from(0 != value % divisor)
-    }
-
-    #[inline]
-    pub fn collect_bool<F: FnMut(usize) -> bool>(len: usize, mut f: F) -> Vec<u64> {
-        let mut buffer = Vec::with_capacity(ceil(len, 64) * 8);
-
-        let chunks = len / 64;
-        let remainder = len % 64;
-        for chunk in 0..chunks {
-            let mut packed = 0;
-            for bit_idx in 0..64 {
-                let i = bit_idx + chunk * 64;
-                packed |= u64::from(f(i)) << bit_idx;
-            }
-
-            // SAFETY: Already allocated sufficient capacity
-            buffer.push(packed);
-        }
-
-        if remainder != 0 {
-            let mut packed = 0;
-            for bit_idx in 0..remainder {
-                let i = bit_idx + chunks * 64;
-                packed |= u64::from(f(i)) << bit_idx;
-            }
-
-            // SAFETY: Already allocated sufficient capacity
-            buffer.push(packed);
-        }
-
-        buffer.truncate(ceil(len, 8));
-        buffer
-    }
 }
